优化框架，实现会员管理

+++ src/baseapps/party/src/cn/gorun8/easyfk/party/controller/PartyClassificationGroupController.java
@@ -15,16 +15,17 @@ package cn.gorun8.easyfk.party.controller;
 
 import cn.gorun8.easyfk.base.util.UtilHttp;
 import cn.gorun8.easyfk.base.util.UtilMessages;
+import cn.gorun8.easyfk.base.util.UtilMisc;
+import cn.gorun8.easyfk.base.util.UtilValidate;
 import cn.gorun8.easyfk.entity.GenericValue;
-import cn.gorun8.easyfk.entity.SequenceFactory;
-import cn.gorun8.easyfk.entity.page.UtilPage;
 import cn.gorun8.easyfk.party.service.PartyClsGroupService;
+import net.sf.json.JSONObject;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.support.ResourceBundleMessageSource;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
-import org.springframework.web.servlet.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -41,20 +42,55 @@ public class PartyClassificationGroupController {
 	
 	@Autowired
 	private PartyClsGroupService partyClsGroupService;
-
+	//ResourceBundleMessageSource t;
 	@RequestMapping(value = "list")
-	public String list( ModelAndView mv ,HttpServletRequest request,@RequestParam(value="pageSize" ,defaultValue="10")int pageSize,
-						@RequestParam(value="pageIndex" ,defaultValue="1")int pageIndex ){
-		UtilPage.startPage(request, pageIndex, pageSize);
-		List<GenericValue> partyClsGroupList = partyClsGroupService.findPartyClsGroupList();
-		mv.addObject("partyClsGroupList",partyClsGroupList);
-		return  "page/partyclsgrouplist";
+	public String list(HttpServletRequest request){
+		String parentId = "N/A";
+		List<Map> parClsMapList = partyClsGroupService.listRootNode(UtilMisc.toMap("parentId",parentId));
+		if(parClsMapList.size() >0) {
+			request.setAttribute("partyClsGroupRoot", parClsMapList);
+			Map<String, Object> attrMap = UtilHttp.getJSONAttributeMap(request, "partyClsGroupRoot");
+			JSONObject json = JSONObject.fromObject(attrMap);
+			request.setAttribute("partyClsGroupList", json);
+			String id = (String)parClsMapList.get(0).get("partyClassificationGroupId");
+			request.setAttribute("partyClsGroupRootId", id);
+			return "page/partyclsgrouplist";
+		}
+
+		return "page/initpartyclsgroup";
+	}
+
+	@RequestMapping(value = "listchild")
+	@ResponseBody
+	public String listchild(HttpServletRequest request,@RequestParam(value="parentId")String parentId){
+		if(UtilValidate.isEmpty(parentId))
+		{
+			return UtilMessages.errorResponse(request,"没有参数");
+		}
+		List<Map> parClsMapList =  partyClsGroupService.listChildNode(UtilMisc.toMap("parentId",parentId));
+		request.setAttribute("partyClsGroupList", parClsMapList);
+		return  UtilMessages.successResponse("partyClsGroupList",request);
+	}
+
+	@RequestMapping("createform")
+	public String createPartyClsGroupForm( HttpServletRequest request,HttpServletResponse response ){
+		createPartyClsGroup(request,response);
+		return "redirect:list";
 	}
 
 	@RequestMapping("create")
 	@ResponseBody
 	public String createPartyClsGroup( HttpServletRequest request,HttpServletResponse response ){
 		Map<String,Object> params = UtilHttp.getParameterMap(request);
+		String parentGroupId = (String)params.get("parentGroupId");
+		String description = (String)params.get("description");
+
+		boolean exits = partyClsGroupService.hasPartyClsGroupName(parentGroupId, description);
+		if(exits)
+		{
+			return  UtilMessages.errorResponse(request, "相同目录下已存在相同名称的机构");
+		}
+
 		try {
 			GenericValue partyClsGroup = GenericValue.fromMap(params);
 			String id = partyClsGroup.newPrimaryKey("partyClassificationGroup", "partyClassificationGroupId");
@@ -71,6 +107,15 @@ public class PartyClassificationGroupController {
 	public String savePartyClsGroup(HttpServletRequest request,HttpServletResponse response ){
 		try {
 			Map<String,Object> params = UtilHttp.getParameterMap(request);
+			String parentGroupId = (String)params.get("parentGroupId");
+			String description = (String)params.get("description");
+
+			boolean exits = partyClsGroupService.hasPartyClsGroupName(parentGroupId, description);
+			if(exits)
+			{
+				return  UtilMessages.errorResponse(request, "相同目录下已存在相同名称的机构");
+			}
+
 			GenericValue partyClsGroup = GenericValue.fromMap(params);
 			partyClsGroupService.savePartyClsGroup(partyClsGroup);
 		}catch (Exception e){
@@ -84,12 +129,17 @@ public class PartyClassificationGroupController {
 	@ResponseBody
 	public String removePartyClsGroupByPrimaryKey(HttpServletRequest request,@RequestParam("partyClsGroupId") String partyClsGroupId){
 		try {
+			boolean rel = partyClsGroupService.hasChildrenPartyClsGroup(partyClsGroupId);
+			if(rel)
+			{
+				return  UtilMessages.errorResponse(request, "存在下级组织机构，不能删除");
+			}
 			partyClsGroupService.removePartyClsGroupByPrimaryKey(partyClsGroupId);
 		}catch (Exception e){
 			e.printStackTrace();
-			return  UtilMessages.errorResponse(request, "更新组织机构出错");
+			return  UtilMessages.errorResponse(request, "删除组织机构出错");
 		}
-		return  UtilMessages.successResponse(request,"更新组织机构成功");
+		return  UtilMessages.successResponse(request,"删除组织机构成功");
 	}
 
 }
